requirements.txt
# Deep Learning
tensorflow==2.15.0
keras==2.15.0
# Data Processing
pandas>=2.1.0
numpy<2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0.1

# Utilities
tensorboard==2.15.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
matplotlib>=3.8.0

# Visualization (optional but recommended for TensorBoard)
h5py>=3.10.0
typing-extensions>=4.8.0